requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
ctrader_open_api==0.9.2
daphne==4.1.2
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docker==7.1.0
eventlet==0.40.0
frozenlist==1.5.0
greenlet==3.2.3
gevent==24.2.1
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inputimeout==1.0.4
kombu==5.5.3
msgpack==1.1.0
multidict==6.4.3
numpy==1.26.4
oandapyV20==0.7.2
pandas==2.2.3
pandas_ta==0.3.14b0
pika==1.3.2
pip==25.0.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==3.20.1
psycopg2-binary==2.9.10
pydantic==2.8.2
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
service-identity==24.2.0
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
tradingview-datafeed==2.1.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
yarl==1.19.0
zope.interface==7.2
whitenoise==6.2.0
django-environ
django-celery-results
django-celery-beat
aio-pika
httpx==0.27.2
tenacity==9.0.0
pybreaker==1.2.0
